refactor : Company 삭제 기능 분리 작업 (#98)
